api/requirements.txt
aiofile==3.7.4
aiofiles==0.8.0
aioredis==2.0.1
alembic==1.7.7
anyio==3.5.0
asgi-lifespan==1.0.1
asgiref==3.5.0
astroid==2.11.3
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
Authlib==1.0.1
backports.cached-property==1.0.1
bcrypt==3.2.0
black==22.3.0
caio==0.9.5
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
Deprecated==1.2.13
deprecation==2.1.0
dill==0.3.4
distlib==0.3.4
docker==5.0.3
dynaconf==3.1.8
ecdsa==0.17.0
fastapi==0.75.2
filelock==3.6.0
graphql-core==3.2.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.22.0
identify==2.5.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.950
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.18.1
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==3.0.8
pytest==7.1.2
pytest-asyncio==0.18.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-i18n==0.3.9
python-multipart==0.0.5
PyYAML==6.0
redis==4.2.2
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
sqlalchemy2-stubs==0.0.2a22
starlette==0.17.1
strawberry-graphql==0.109.0
testcontainers==3.5.3
toml==0.10.2
tomli==2.0.1
types-aiofiles==0.8.8
types-certifi==2021.10.8.2
types-python-dateutil==2.8.15
types-six==1.16.15
types-ujson==4.2.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.14.1
watchgod==0.8.2
websocket-client==1.3.2
websockets==10.3
wrapt==1.14.0
